add find all authors with pagination add search index to author schema check books availability before deleting author

+++ src/repositories/AuthorRepository.js
@@ -1,8 +1,41 @@
 const Author = require('../models/AuthorSchema');
+const UnprocessableError = require("../errors/UnprocessableError");
 
 const findAllAuthors = async () => {
     try {
         return await Author.find();
+    } catch (error) {
+        throw error;
+    }
+}
+
+const findAllAuthorsWithPagination = async (page, size) => {
+    try {
+        const totalCount = await Author.countDocuments();
+        const totalPages = Math.ceil(totalCount / size);
+        const skip = (page - 1) * size;
+        const from = skip + 1;
+
+        const authors = await Author.find({}).skip(skip).limit(size);
+        const to = skip + authors.length;
+
+        return {authors, totalCount, totalPages, from, to};
+    } catch (error) {
+        throw error;
+    }
+}
+
+const findAllAuthorsBySearchWithPagination = async (searchText, page, size) => {
+    try {
+        const totalCount = await Author.find({$text: {$search: searchText}}).countDocuments();
+        const totalPages = Math.ceil(totalCount / size);
+        const skip = (page - 1) * size;
+        const from = skip + 1;
+
+        const authors = await Author.find({$text: {$search: searchText}}).skip(skip).limit(size);
+        const to = skip + authors.length;
+
+        return {authors, totalCount, totalPages, from, to};
     } catch (error) {
         throw error;
     }
@@ -38,6 +71,11 @@ const updateAuthor = async (params, authorData) => {
 
 const deleteAuthor = async (params) => {
     try {
+        const author = await Author.findById(params.id);
+
+        if (author.books.length) {
+            throw new UnprocessableError('Could not delete. This author associated with books!');
+        }
         return await Author.findByIdAndDelete(params.id);
     } catch (error) {
         throw error;
@@ -45,5 +83,11 @@ const deleteAuthor = async (params) => {
 }
 
 module.exports = {
-    findAllAuthors, createAuthor, findAuthorById, updateAuthor, deleteAuthor
+    findAllAuthors,
+    findAllAuthorsWithPagination,
+    findAllAuthorsBySearchWithPagination,
+    createAuthor,
+    findAuthorById,
+    updateAuthor,
+    deleteAuthor
 }
